+++ rts/motoko-rts/src/stabilization/graph_copy/time.rs
@@ -2,22 +2,22 @@
 /// Deterministically measured in synthetic steps.
 // TODO: Unify with the bounded time used in the incremental GC.
 pub struct BoundedTime {
-    steps: usize,
-    limit: usize,
+    steps: u64,
+    limit: u64,
 }
 
 impl BoundedTime {
-    pub fn new(limit: usize) -> BoundedTime {
+    pub fn new(limit: u64) -> BoundedTime {
         BoundedTime { steps: 0, limit }
     }
 
     pub fn tick(&mut self) {
-        self.steps = usize::saturating_add(self.steps, 1);
+        self.steps = u64::saturating_add(self.steps, 1);
     }
 
-    pub fn advance(&mut self, amount: usize) {
+    pub fn advance(&mut self, amount: u64) {
         debug_assert!(amount > 0);
-        self.steps = usize::saturating_add(self.steps, amount);
+        self.steps = u64::saturating_add(self.steps, amount);
     }
 
     pub fn is_over(&self) -> bool {
